pyproject.toml
[project]
name = "dgsp"
version = "0.1.0"
description = ""
authors = [{ name = "LumbaBalumba", email = "vlad.lazar.ik528@yandex.ru" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "scipy (>=1.15.2,<2.0.0)",
  "pandas (>=2.2.3,<3.0.0)",
  "scikit-learn (>=1.6.1,<2.0.0)",
  "filterpy (>=1.4.5,<2.0.0)",
  "numba (>=0.61.0,<0.62.0)",
  "sympy (>=1.13.3,<2.0.0)",
  "joblib (>=1.4.2,<2.0.0)",
  "matplotlib (>=3.10.1,<4.0.0)",
  "typing-extensions (>=4.13.1,<5.0.0)",
  "cupy-cuda12x (>=13.6.0,<14.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "scripts.generate:generate_all"
estimate = "scripts.estimate:estimate"
stats = "scripts.stats:stats"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

